test: move DetectLatestUpstreamOfficialTag func
Move DetectLatestUpstreamOfficialTag from utils
to operator_test to make utils shorter.
Also move getTagHint that is used by DetectLatestUpstreamOfficialTag.
