+++ tests/BUILD.bazel
@@ -38,8 +38,6 @@ go_library(
         "//tests/libvmi:go_default_library",
         "//tests/testsuite:go_default_library",
         "//tests/util:go_default_library",
-        "//vendor/github.com/Masterminds/semver:go_default_library",
-        "//vendor/github.com/google/go-github/v32/github:go_default_library",
         "//vendor/github.com/google/goexpect:go_default_library",
         "//vendor/github.com/k8snetworkplumbingwg/network-attachment-definition-client/pkg/apis/k8s.cni.cncf.io/v1:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
@@ -233,6 +231,7 @@ go_test(
         "//tests/util:go_default_library",
         "//tests/virtctl:go_default_library",
         "//tools/vms-generator/utils:go_default_library",
+        "//vendor/github.com/Masterminds/semver:go_default_library",
         "//vendor/github.com/evanphx/json-patch:go_default_library",
         "//vendor/github.com/google/go-github/v32/github:go_default_library",
         "//vendor/github.com/google/goexpect:go_default_library",
